client - rename files in legacy folder
