+++ packages/js/client/src/legacy/types/GenericIUriPackage.ts
