+++ packages/js/client/src/legacy/types/GenericIUriRedirect.ts
