+++ packages/js/client/src/legacy/types/GenericIUriWrapper.ts
