+++ packages/js/client/src/legacy/types/GenericInterfaceImplementations.ts
