+++ packages/js/client/src/legacy/types/index.ts
@@ -1,6 +1,6 @@
-export * from "./InterfaceImplementations";
-export * from "./Env";
-export * from "./IUriWrapper";
-export * from "./IUriRedirect";
-export * from "./IUriPackage";
-export * from "./UriResolverLike";
+export * from "./GenericInterfaceImplementations";
+export * from "./GenericEnv";
+export * from "./GenericIUriWrapper";
+export * from "./GenericIUriRedirect";
+export * from "./GenericIUriPackage";
+export * from "./GenericUriResolverLike";
